==> scripts/00_DataWrangling.ipynb <==
# Pan-barley transcriptome data wrangling
## 1. Subset 1:1x20 Orthogroups
## 2. Create metatable for all accessions
## 3. Divide TPM table into accession-sets

library(tidyverse)
# import tables 
getwd()
tpm <- read.table("PanBaRT20_tpm_genes.csv", sep = ",", row.names = 1, header = TRUE, stringsAsFactors = TRUE)
hogs <- read.table("GsRTD_and_PanBaRT20_match.tsv", sep = "\t", header = TRUE, stringsAsFactors = TRUE)
head(tpm, 2)
dim(tpm)
dim(hogs)
head(hogs, 2)

## 1. Subset 1:1x20 Orthogroups

orts <- hogs %>% filter(Pan_category == "core;single") %>% 
                 select(PanBaRT20_gene_id, PanBaRT20_transcript_id, GsRTD_gene_id, Genotype) %>% 
                 distinct()
glimpse(orts)
singlecopy <- orts %>% select(PanBaRT20_gene_id) %>% distinct()
glimpse(singlecopy)
# Save
getwd()
write.csv(singlecopy, file = "Singlecopy_PanBaRT20gene_all.csv", row.names = FALSE)

# Subset tpm table with 20let-HOGs translated to PanBaRT20_gene_id-s
tpm_geneID <- tpm %>% rownames_to_column("PanBaRT20_gene_id")
tpm_filt <- inner_join(singlecopy, tpm_geneID, by = "PanBaRT20_gene_id") %>% 
            column_to_rownames("PanBaRT20_gene_id")
dim(tpm_filt)
head(tpm_filt, 2)

trans_tpm <- as.data.frame(t(tpm))
head(trans_tpm, 2)

## 2. Create metatable for all accessions

# Create metatable
##################
meta <- trans_tpm %>% rownames_to_column("ID") %>%
                        select(ID) %>%
                        mutate(Accession = str_extract(ID, regex("[a-z0-9]+", ignore_case = TRUE))) %>%
                        mutate(Tissue = str_sub(str_extract(ID, regex("_[a-z]+", ignore_case = TRUE)), -2)) %>%
                        mutate(Batch = as.integer(str_extract(ID, "(?<=_\\w{2})\\d$"))) %>%
                        mutate(Tissue = case_when(Tissue == "Ca" ~ "Caryopsis",
                                                  Tissue == "Co" ~ "Coleoptiles",
                                                  Tissue == "Sh" ~ "Shoot",
                                                  Tissue == "Ro" ~ "Root",
                                                  Tissue == "In" ~ "Inflorescence"))
head(meta, 2)
glimpse(meta)

# save joined metatable for all accessions together
getwd()
write.csv(meta, file = "PanBaRT20_geneTPM_meta.csv", row.names = FALSE)

# save project table with each accession mentioned once
meta_acc <- meta %>% select(Accession) %>% distinct()
dim(meta_acc)
write.table(meta_acc, file = "PanBaRT20_project_table.txt", row.names = FALSE)

# save accession-wise metatable
acc <- unique(meta$Accession)
for (ac in acc) {
    # For the special case of ZDM01467: filter out In2 sample
    if (ac == "ZDM01467") {
        subset_meta <- meta %>% 
            filter(ID != "ZDM01467_In2") %>% 
            filter(Accession == ac)
    } else {
        subset_meta <- meta %>% filter(Accession == ac)
    }
    filename <- paste0(ac, "_meta.csv")
    write.csv(subset_meta, file = filename, row.names = FALSE)
}

## 3. Divide TPM table into accession-sets

# divide accessions into separate tpm tables and save them
getwd()
# create vector of accession types
acc <- unique(meta$Accession)
acc

# First, determine the set of genes that pass the filter across all accessions
common_filtered_genes <- Reduce(intersect, 
                                lapply(acc, function(name) {
                                    matching_cols <- grep(paste0("^", name), colnames(tpm_filt))
                                    subtable <- tpm_filt[, matching_cols, drop = FALSE]
                                    
                                    # Special case for "ZDM01467" to exclude "ZDM01467_In2" column
                                    if (name == "ZDM01467") {
                                        subtable <- subtable[, !colnames(subtable) %in% c("ZDM01467_In2"), drop = FALSE]
                                    }  
                                    
                                    idx <- rowSums(subtable >= 0.5) >= 2
                                    return(rownames(subtable)[idx])
                                }))

# Save the common_filtered_genes list to a CSV file
common <- data.frame(PanBaRT20_gene_id=common_filtered_genes)
common_single20 <- inner_join(common, orts, by = "PanBaRT20_gene_id") %>% 
                    select(PanBaRT20_gene_id, PanBaRT20_transcript_id, Genotype) %>%
                    distinct()
glimpse(common_single20)
write.csv(common_single20, file = "Singlecopy_transcript_to_gene_expressing.csv", row.names = FALSE, quote = FALSE)
only_geneid <-common_single20 %>% 
                select(PanBaRT20_gene_id) %>% 
                distinct()
write.csv(only_geneid, file = "Singlecopy_PanBaRT20geneID_expressing.csv", row.names = FALSE, quote = FALSE)

getwd()

# Extract subtables for matching column names
subtables <- lapply(acc, function(name) {
  matching_cols <- grep(paste0("^", name), colnames(tpm_filt))
  subtable <- tpm_filt[rownames(tpm_filt) %in% common_filtered_genes, matching_cols, drop = FALSE]
  
  # Special case for "ZDM01467" to exclude "ZDM01467_In2" column
  if (name == "ZDM01467") {
    subtable <- subtable[, !colnames(subtable) %in% c("ZDM01467_In2"), drop = FALSE]
  }  
  
  print(paste0(name, " filtered genes with >0.5 for min 2 samples across all samples:", dim(subtable)))
  
  # Save each subtable in a separate file
  file_name <- paste0("PanBaRT20_geneTPM_ort_filt_", name, ".csv")
  write.csv(subtable, file = file_name, row.names = TRUE)
})

sessionInfo()
