==> scripts/01_varstabDESeq2_pgsblaptop.R <==
# script for deseq2 varstab
library("tidyverse")
library("DESeq2")
set.seed(42)
# import filtered count tables and metatables
setwd("/home/vanda/Documents/PanBarley_transcriptome/00_geneTPM_peraccession/")
fnames <- list.files(pattern = "PanBaRT20_geneTPM_ort_filt_*")
cntlist = list()
for (i in fnames) {
    counts <- read.table(i, row.names = 1, header = TRUE, sep = ",", stringsAsFactors = FALSE)
    cnt <- counts %>% mutate_if(is.numeric, as.integer)
    cntlist[[i]] <- cnt
}
setwd("/home/vanda/Documents/PanBarley_transcriptome/00_meta_peraccession/")
fnames <- list.files(pattern = "*_meta.csv")
metalist = list()
for (i in fnames) {
    m <- read.table(i, row.names = 1, header = TRUE, sep = ",", stringsAsFactors = FALSE)
    #cnt <- cnt %>% mutate_if(is.numeric, as.integer)
    metalist[[i]] <- m
}
#
setwd("/home/vanda/Documents/PanBarley_transcriptome/01_vsdTPMs_peraccession/")
# 
for (i in 1:20) {
    if (all(colnames(cntlist[[i]]) == rownames(metalist[[i]]))) {
        # create design
        dds <- DESeqDataSetFromMatrix(countData = cntlist[[i]],
                                   colData = metalist[[i]],
                                   design = ~ Tissue)
        dds <- estimateSizeFactors(dds)
        print(summary(sizeFactors(dds)))
        dds <- estimateDispersions(dds)
        vsd <- varianceStabilizingTransformation(dds, blind = FALSE)
        vsdMat <- assay(vsd)
        print(dim(vsdMat))
        #print(head(cntlist[[i]], 1))
        #print(head(metalist[[i]], 3))
        write.table(vsdMat, 
                    file = paste0(metalist[[i]]$Accession[1], "_vsdTPM.csv"), 
                    append = FALSE, 
                    quote = FALSE, 
                    sep = ",", 
                    dec = ".", 
                    row.names = TRUE, 
                    col.names = TRUE)
        }
}
sessionInfo()
